backend/requirements.txt
# requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
newspaper3k==0.2.8
python-dotenv==1.0.0
openai==0.28.0
requests==2.31.0